Have separate fields for volume, long file name, short file name, and Unicode file name offsets.
Display the "UNIX privileges" field as a 4-longword structure, as part
of the AFP 3.0 spec claims; unfortunately, another part claims it's only
4 bytes, but, unless a "long" is 8 bits, one or the other claim must be
false.

Completely get rid of the static "afp_handle" variable - it's not
necessary.

Put the Unicode name offset into the tree, if present, in
"parse_file_bitmap()" and "parse_dir_bitmap()".

Get rid of some unused variables.

Put the newline mask and character into the tree for FPRead.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5289 f5534014-38df-0310-8fa8-9805f1628bb7