Added a routine to dissect a DOS error code.  These are used in the spoolss (also registry and srvsvc?) pipe dissectors.
Also added some helper routines for decoding a range of uint8s.  These map
nicely to the dissect_{dcerpc,ndr}_foo() format with the addition of a
length parameter, but aren't really part of the NDR specification and are
probably only going to be used in the NT dcerpc dissectors.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5421 f5534014-38df-0310-8fa8-9805f1628bb7