NCP-over-UDP doesn't have a special header, as NCP-over-TCP does.
Dissect NCP watchdog packets.

Clean up the handling of NCP packet-specific headers and payload a bit.

Put in a protocol tree item for the upper 8 bits of the connection
number.

Make the "sorry, we don't dissect this yet" protocol tree item cover the
part we *don't* dissect, rather than covering the packet type.

Use the data dissector to dissect the data in a burst mode packet.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5478 f5534014-38df-0310-8fa8-9805f1628bb7