Add LANE 2.0 stuff.
Show the maximum frame size symbolically.

For the various LANE frame types, show only the fields those frame types
are specified to contain, and show only the flag bits that are specified
for those frame types.

Use the ELAN name size when putting the name into the protocol tree, and
don't put it in if the name size is 0.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5543 f5534014-38df-0310-8fa8-9805f1628bb7