From Graeme Hewson:
	Allow "-" as the output file name in Wiretap, referring to the
	standard error.

	Optimize the capture loop.

Fix some of the error-message printing code in Ethereal and Tethereal.

Have Wiretap check whether it can seek on a file descriptor, and pass
the results of that test to the file-type-specific "open for output"
routine.  Have the "open for output" routines for files where we need to
seek when writing the file return an error if seeks don't work.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5884 f5534014-38df-0310-8fa8-9805f1628bb7