Get rid of some "unused variable" and "unused static function" warnings, and some compile errors in Tethereal, when compiling without libpcap.
If libpcap is missing (whether that's detected at compile time or, as on
Windows, at run time), don't call any of the "capture_prefs" routines -
the routine to create the capture preferences page wasn't called, so the
other routines can try to refer to non-existent widgets and other items
and crash.

Get rid of the stub routines in "capture_prefs()" used when compiling
without libpcap, as they're no longer called.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5888 f5534014-38df-0310-8fa8-9805f1628bb7