If the length of a string is zero, don't put it into the protocol tree.
Handle the case where "get_unicode_or_ascii_string()" returns a null
pointer (which can be the case if the length supplied is zero, which we
check for as per the above, but can also be the case for a Unicode
string if the length supplied is 1 byte).

Fix a call to "proto_tree_add_uint()" that was presumably supposed to be
a call to "proto_tree_add_item()".


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6015 f5534014-38df-0310-8fa8-9805f1628bb7