From Ulf Lamping:
  Generalize paths and get rid of sed-dependent ethereal.nsi.in.

Remove ethereal.nsi.in, and re-add ethereal.nsi with Ulf's changes.
Modify Makefile.nmake to use it.

Note:  You'll have to add "COMMON_FILES_GNU" path to your config.nmake
that points to a directory with all of Ethereal's required DLLs.

Be obnixious and set "SetShellVarContext all" so that we install for
every user, like most standard Windows apps.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6020 f5534014-38df-0310-8fa8-9805f1628bb7