From Andy Hood: in the AODV6 dissector, rename "ext_t" to "aodv6_ext_t", as "ext_t" is defined by a system header file on AIX.
Get rid of the bitfields in "struct aodv_rrep" and "rrep_t", as some
compilers don't allow any type other than "int" or "unsigned int" for
bitfields (so "guint8" won't work), and as there's no guarantee of the
order of bitfields in a structure.

Make "rreq", "rrep", and "rerr" local to "dissect_aodv()" and
"dissect_aodv6()", as they're not used outside them.

Add a protocol tree item for the prefix size field in AODV and AODV6 replies.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6060 f5534014-38df-0310-8fa8-9805f1628bb7