Add a "gssapi_lookup_oid()" that takes a binary OID (pointer and length) as an argument, and looks up that OID in the GSSAPI OID hash table.
Always use that routine to look up OIDs, so that we never use the result
of "format_oid()" as the key (as that doesn't necessarily work).

Make "gssapi_oids" static, as one should only look up GSSAPI
authentication mechanism OIDs with "gssapi_lookup_oid()".

In the SPNEGO dissector, free up the OID strings when we're done with
them, and don't advance the offset past the OID until after we put the
OID into the protocol tree.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6228 f5534014-38df-0310-8fa8-9805f1628bb7