Get rid of extra blanks at ends of program names (they're not needed; in fact, they cause an extra blank to appear).
Turn "get_string()" into a routine that fetches the length of a string,
fetches the string, puts the string into the tree with the offset of the
length as the starting offset, and returns the new offset.

Make routines that return offsets return an "int", as packet offset
variables are "int"s.

Fix up a bunch of "proto_tree_add_uint()" calls, including presumably
properly updating the offset after the call.

Clean up the handling of qualified names, integer-or-OID types, and even
objects a bit.

Put octet strings into the protocol tree with the offset of the length
as the starting offset.

Fix up the record length variable to be decimal, and just use
"proto_tree_add_uint()" to put it into the tree.

In at least one capture I've seen, there are only two bytes after the
NDPS tree field in a Bind PA request, not 4, even if the NDPS tree name
isn't an empty string.

FT_BYTES and FT_BOOLEAN don't have a base for display, so use BASE_NONE.

Fix the field for address bytes to have the right name (it's not just
the length, it's the length *and* the bytes).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6467 f5534014-38df-0310-8fa8-9805f1628bb7