Note why we're including <sys/socket.h> and <winsock2.h> in places where we're including them to get AF_ values.
Shuffle the order of those includes in "inet_pton.c" to match the order
in "inet_ntop.c".

Include <winsock2.h> in "resolv.c" and "to_str.c" so that we get AF_
values from it - including AF_INET6 - if they exist, as we want to have
everybody who uses AF_INET6 get the same value.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6517 f5534014-38df-0310-8fa8-9805f1628bb7