Values in Microsoft protocols that don't have their byte order specified by DCE RPC are usually little-endian; fix a bunch of "proto_tree_add_item()" calls (most are for byte-array or string fields, so the byte order doesn't make a difference, but one is a number).
Put an item into the protocol tree for the encrypted NT password block.

Mallocate the buffer for the Unicode version of the password, rather
than assuming it'll fit in 256 bytes.

"g_malloc()" never returns NULL - it either allocates memory or aborts -
so don't check for a mallocation failure.

Don't try to decrypt the NT password block if we don't have a password.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6731 f5534014-38df-0310-8fa8-9805f1628bb7