Fix for DCERPC detection when carried ontop of SMB.
If we do not see the TreeConnect call when a TID is connected, we did not
know it was a IPC share.
If we do not know what kind of share it is we assume it being a normal one
and thus read/write data to that share is normal file i/o.


Update the dissector so that IF it sees a Transaction SMB carrying PIPE (dcerpc)
then we assume that all other read/write to that TID is also DCERPC.
I.e. we assume the entire TID is IPC.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6747 f5534014-38df-0310-8fa8-9805f1628bb7