Rename "Go To Specified Frame" to "Go To Corresponding Frame" (I'm not sure there *is* a good name for it).
Put it in the "Tools" menu as well (although I'm not sure what top-level
menu it belongs in, or if it should get a new one).

Make those items sensitive only if there's an FT_FRAMENUM item selected
in the protocol tree pane.

Clean up some menu paths in "set_menu_sensitivity()" (not that it
matters, as only the last component, and the first component if it's a
menu factory name, are used).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@6803 f5534014-38df-0310-8fa8-9805f1628bb7