Treat the Peripheral Qualifier field the same way we treat the Peripheral Device Type field, i.e. as a registered field that's a bitfield with a value_string table.  Give it the correct bitmask.
Fix up some device type names, and add some additional ones from the 25
July 2002 SPC-3 draft.

Correctly adjust the length of the LUN list in the Report LUNs response
(subtract 8 from the payload length before comparing, don't subtract 8
from the resulting length after comparing and adjusting), and rename the
variable to reflect the fact that it's a length in bytes, not a number
of LUNs.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@7322 f5534014-38df-0310-8fa8-9805f1628bb7