Fixed a small bug in tcp sequence number analysis.
FIN flag would previously only add one to the sequence number if the
FIN packet was empty, i.e. did not carry any payload data.

This caused ethereal to incorrectly flag the ACK to such packets
(FIN+payload data) to be incorrectly flagged as
ACK to previously lost segment.

Change the algorithm to always add 1 to the segment length, and thus the sequence number for all packets with teh FIN bit set.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@7371 f5534014-38df-0310-8fa8-9805f1628bb7