Use "gboolean" for Boolean variables.
Allocate the per-conversation and per-frame data items from a GMemChunk
(which saves memory and CPU time, *and* lets us free all those items
quickly - as opposed to not freeing them at all, as had been the case).

Don't assume that, just because a conversation for the TCP connection
exists, it necessarily has an AJP13 data chunk attached to it; the
conversation might have, for example, been created by the TCP sequence
number analysis code.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@7374 f5534014-38df-0310-8fa8-9805f1628bb7