The MID field in the SMB header is in reality only ~10 bits in size so it is pretty common for MID values to be reused even in moderately sized captures.
The test to compare that the command type between the request
and reply is not sufficient for when most of the commands between the client
and the server are the same (e.g. streaming Read/Write)

Change the matching so that ONLY the first "response" we see for a certain
open MID will be matched to the original request.

I.e.  Prevent
Read Request
Read Reply
[missing from capture] Read Request
Read Reply

From incorrectly matching the second reply (if it has a reused MID) with the
first request.

This makes the response time statistics a bit more reliable as well.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@7888 f5534014-38df-0310-8fa8-9805f1628bb7