Make the default sorting for the list be column 2 in descending order so we get the rocedures with the most number of calls at the top.
Fix a bug added in previous checkin.
When the table can be sorted in different ways than just the procedure number sorted in ascending order, the row for a specific procedure may change everytime we drawi/sort the table.
Add code to keep track of which row a procedure is currently positioned at in the list so that we statistics are added to the correct entry.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@7906 f5534014-38df-0310-8fa8-9805f1628bb7