Updates to the fibre channel dissector
removed two hf fields that are now redundant since F_CTL is decoded

placed most interesting header fields in a struct

added request response matching and measurement of delta time between the first and the last frame of an exchange.

todo:
1, finish restructure of decoding of fc header so that
it will populate the tree as it reads more data fields from the tvb
so that we can get useful info in the tree even for short frames.
2, add TAP to fc and pass the header struct and the time delta struct to
tap listeners
3, add ServiceResponseTime tap listener for FC


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@7923 f5534014-38df-0310-8fa8-9805f1628bb7