Attempt to decode GeneralString. RFC2833 is a bit "unclear" but I guess this type is encoded as first a length-determinant followed by the actual ascii data.
I belive the length-determinant is byte aligned in aligned-per so the entire
field is so.

at best, this is pure guesswork  but it does decode the single capture i do have containing GeneralString types properly.

Anyone interested are welcome to purchase and provide
proper h323 standard docs from itu-t and snail-mail them to me.
A random asn file from www.packetyzer.com together with the X.691 pdf file is
"difficult".


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@7969 f5534014-38df-0310-8fa8-9805f1628bb7