The length and offset for large file support for LockingAndX was broken.
We treated this as just a normal 64bit integer in LittleEndian format.
However, this is actually 2   32 bit integers, each in LittleEndian format
but the two 32 bit fields are stored in BigEndian format relative to each other.

Since we dont do 64 bit aritmetic I had to convert the field to FT_STRING as well
so sorry, no creative len>xxx    filters anymore.   but at least we present
the data in the correct way in the tree pane.

We didnt see this one earlier since most locking_andx requests are probably for offset : 0   and length: -

Funnily enough it seems that certain popular commercial products have the same bug as ethereal had up until 5 minutes ago.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8196 f5534014-38df-0310-8fa8-9805f1628bb7