Update manpage for ethereal. Update the talkers tap for tethereal (iousers) and change the command line to invoke the tethereal version from  -z io,users,   to -z talkers, to be the same as for ethereal.
Sorry if it breaks some scripts but io,users was a very nonintuitive name for this option.
talkers is not much better  but at least a little bit more descriptive/intuitive.   Anyone with a better name for this are welcome to provide a patch.

The tethereal version is now agnostic to wether v4 or v6 are transporting UDP/TCP


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8236 f5534014-38df-0310-8fa8-9805f1628bb7