Updates for the endpoint talkers thing
The table now has a popup menu with
   Match display filter
         Selected
               EP1 <-> EP2
               EP1 --> EP2
               EP1 <-- EP2
               EP1 --> ANY
               EP1 <-- ANY
               EP2 --> ANY
               EP2 <-- ANY
          Not Selected
               ...
          ...
    Prepare Display Filter
          ...

Self explanatory.

Now the bad news.
I set the display filter box in the main window and i call redissect_packet
which redissects the packet list   but the displayfilter does not take
or affect the main window until i click  the apply button.

Some signal needs to be raised to some object me thinks.

Please feel free to fix it if you know what is missing.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8279 f5534014-38df-0310-8fa8-9805f1628bb7