fix to various stats tables.
The code used to rely on min_time==0 to determine whether this was the first
packet or not and whereby we had to initialize min_time to the current value.

This obviously does not work for capture files with poor timestamp resolution
where the response time is actually, according to the capture file, 0
and we got all sorts of weird effects like average response time being less than the minimum response time.

note, the bug only affected the minimum response time in the tables and not max or average response time.
it would "miss" tose minimum response times that were ==0 and display the minumin response time in the capture that were >0


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8358 f5534014-38df-0310-8fa8-9805f1628bb7