For (non-heuristic) SIP-over-TCP, dissect stuff that's neither a request nor a response as continuation data.  For SIP-over-everything-else, reject it.
Parse the headers regardless of whether we're building a protocol tree
or not; if we're not, we just do it to look for a blank line separating
the headers from the body.  Do that instead of scanning for the message
body separately.

When scanning for a colon, don't scan past the end of the line.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8549 f5534014-38df-0310-8fa8-9805f1628bb7