Make the "needs_redraw" flag a gboolean, as it's a Boolean flag.
Rename "max_count_types and "max_calc_types" to "count_type_names" and
"calc_type_names", to make it clearer what they are.

For the advanced statistics, give different error messages for the case
where no field name was specified and where an invalid field name was
specified.

Give better error messages for the cases where the calculation type
isn't supported for a particular field.

Initialize the calculation type for a given field.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8630 f5534014-38df-0310-8fa8-9805f1628bb7