Make the window geometry information static - it's not used outside this file.
Fetch the geometry information whenever we get a configure_notify event,
i.e. if it actually changes, rather than doing so when we get asked to
delete the main window or we exit.  Don't save the geometry if we've
never gotten such an event, as it presumably means the size and position
haven't changed.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8634 f5534014-38df-0310-8fa8-9805f1628bb7