"add_content_type()" returns either a numerical content type and a null content type pointer or a 0 numerical content type and a non-null content type pointer; if the content type is numerical, we cannot match on the string content type (as we'd dereference a null pointer) and, if the content type is a string, we should not match on the numerical content type (as the value is bogus).
We also have to call "add_content_type()" to get the content type
regardless of whether we're building a protocol tree, as we need to call
subdissectors regardless of whether we're building a protocol tree.

We also need to set the columns regardless of whether we're building a
protocol tree.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8675 f5534014-38df-0310-8fa8-9805f1628bb7