Don't try to get the list of link-layer header types for an interface if the interface isn't in the list of known (local) interfaces - that way we don't try to get entries if the user's in the middle of changing the "Interface:" text, or if the interface is an rpcap: URL (getting the link-layer header type list for a remote interface is currently impossible, and even just getting its *default* interface type could hang for a long time if the remote machine isn't responding).
Free the link-layer header type list when we're done with it.

Label the option menu for that list "Link-layer header type", as it
doesn't control the data link type of the interface in the only case I
know of where it's settable, namely 802.11 interfaces - it just controls
whether the packets you get from the interface have a fake Ethernet
header or a real 802.11 header.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8867 f5534014-38df-0310-8fa8-9805f1628bb7