Bugfix for ethereal crashes
If the ByteCount field in the SMB PDU spanned beyond the end of the
packet because the packet was short or because the BC field was corrupted and contained
garbade data  then the tree item for the command (the subtree just after the SMBHeader subtree) would describe data continuing beyond the end of the
packet.

If we selected one such tree in the dissect pane and used Prepare/Match Selected  this would cause the filter build thing to try to access data beyod the end
of the packet and ethereal would dump core.

Change the END_OF_SMB macro so that it shrinks bc so that bc never describes
data beyond the end of the packet.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8926 f5534014-38df-0310-8fa8-9805f1628bb7