updates to ip_to_str[_buf]() to make them faster.
This functin was in a recent profiling run the fifth most expensive function.
It consumed ~3.5% of the total cpu in that run.

Using a static table this function is now just over twice as fast as previously.

My testrun calling it 10.000.000 times in a loop changed the execution time from ~17.3s ro ~8.2s

If this test was representative for the speedup and if ~3.5 in the profile run
was representative for normal cases this should make ethereal ~1.8% faster.
Woohoo.
(low hanging fruit)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@8992 f5534014-38df-0310-8fa8-9805f1628bb7