Add an entry in the byte_view pane that allows the user to save the highlighted data to a file. This allows the user to select some stuff, and analyse it with external tools, and is very useful for quickly prototying dissectors etc.
This works by retrieving the info that is needed to define where the selected
region is.

It puts up a dialog box that asks for the file to save in.

However, it is an ugly hack, because it reuses print_file_cb, while
print_file_cb should be moved into file_dlg.c.

It also needs to have some warning dialogs put up in error cases.

Finally, it would be good to be able to select a region in the byte_view,
which you can do with click and drag, and then have a menu item to save the
selected bytes.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@9116 f5534014-38df-0310-8fa8-9805f1628bb7